Implement passing of additional log entry data (#6)
Additional data can be set as logrus fields on the entry and are now passed through to the modified JSON payload.
In order to do so the `message` field was moved to jsonPayload, as specified in the [docs](https://cloud.google.com/logging/docs/structured-logging#structured_logging_special_fields).

Fixes #5.

Signed-off-by: Alex Klinkert <alex@klinkert.io>